project-3-advanced/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.20
pydantic==2.10.4
chromadb==0.5.23
faiss-cpu==1.9.0
openai==1.58.1
numpy==2.2.1
python-dotenv==1.0.1
